StkFrames::operator+:  cast nFrames to unsigned int given we add it to a value of type unsigned int

+++ include/Stk.h
@@ -495,7 +495,7 @@ inline StkFrames StkFrames::operator+(const StkFrames &f) const
     Stk::handleError( error.str(), StkError::MEMORY_ACCESS );
   }
 #endif
-  StkFrames sum(nFrames_,nChannels_);
+  StkFrames sum((unsigned int)nFrames_,nChannels_);
   StkFloat *sumPtr = &sum[0];
   const StkFloat *fptr = f.data_;
   const StkFloat *dPtr = data_;
